unit tests with RPS; next up, integration tests
